refactor: change type structure
