+++ src/firebase/firestore/checkRegistration.ts
@@ -10,7 +10,7 @@ import { fetchUserItemRefs } from "@/firebase/firestore/item";
 import { DocumentData } from "firebase-admin/firestore";
 import { DuplicateItemError } from "@/firebase/firestore/errors";
 
-import { ItemParams } from "@/utils/type";
+import { ItemParams } from "@/utils/types/items";
 import { db } from "@/firebase/init";
 
 export class RegistrationValidator {
