+++ src/utils/types/tags.ts
@@ -1,0 +1,6 @@
+import { DocumentReference } from "firebase/firestore";
+
+export type TagData = {
+  tagName: string;
+  itemIds: DocumentReference[];
+};
